Add exception lists for known missing areas/GSS codes
When we run these two scripts, we repeatedly output the same list
of known areas that have now got a different GSS code, or have been
merged with another council/area, and is not always clear whether
these are known updates. Having these lists should make it easier
to spot new exceptions with new data, and we can update them as
needed.

The syntax for printing strings has also changes for `Python3` so
have also updated this for these scripts. They are our own custom
scripts (not forked from Mysociety).